Miscellaneous EIP-2930 Updates (#1897)
* Update Signature hash preimage
* Update json rpc to list fields for access list
* Update what is considered "raw" hash

+++ plugin-api/src/main/java/org/hyperledger/besu/plugin/data/TransactionType.java
@@ -27,8 +27,8 @@ public enum TransactionType {
     this.typeValue = typeValue;
   }
 
-  public int getSerializedType() {
-    return this.typeValue;
+  public byte getSerializedType() {
+    return (byte) this.typeValue;
   }
 
   public static TransactionType of(final int serializedTypeValue) {
